+++ packages/workspace/src/index.ts
@@ -3,4 +3,5 @@
  * @since 0.2.0-alpha.1
  */
 
+export * from "./FrameworkCapabilityDirection";
 export * from "./FrameworkReadiness";